+++ src/core/client/startup.ts
@@ -74,6 +74,7 @@ import './systems/defaults/toolbar';
 
 // Systems
 import './systems/acceptDeclineEvent';
+import './systems/adminControl';
 import './systems/animations';
 import './systems/arrest';
 import './systems/athenaEvents';