+++ src/core/server/imports/systems.ts
@@ -1,4 +1,5 @@
 import '../systems/account';
+import '../systems/adminControl';
 import '../systems/global';
 import '../systems/identifier';
 import '../systems/inventory/clothing';